C Online Learning Platform (Group Project)
Java EE and MySql are used in this platform. It is an online learning
website where learners can get materials to learn C programming
language. They can also take programming tests and upload their code to
the website. The code will be compiled and the results will be responded
to the learners.